Fighting the lint still

+++ malariagen_data/anoph/hap_freq.py
@@ -96,7 +96,7 @@ class AnophelesHapFrequencyAnalysis(
         cohorts_iterator = self._progress(
             coh_dict.items(), desc="Compute allele frequencies"
         )
-        hap_track: dict[int, int] = {}
+        hap_track: dict[np.int64, int] = {}
         for coh, loc_coh in cohorts_iterator:
             hap_track = {k: 0 for k in hap_track.keys()}
             n_samples = np.count_nonzero(loc_coh)
